website/requirements.txt
# imagic website — additional dependencies
# Install alongside the main project: pip install -r website/requirements.txt

fastapi>=0.104
uvicorn[standard]>=0.24
python-multipart>=0.0.6
jinja2>=3.1
python-dotenv>=1.0
authlib>=1.3
itsdangerous>=2.2
httpx>=0.27
stripe>=8.0
